Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-canvas-element/security.pattern.image.strokeStyle-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
security.pattern.image.strokeStyle
Setting strokeStyle to a pattern of a different-origin image makes the canvas origin-unclean
Actual output:

PASS Setting strokeStyle to a pattern of a different-origin image makes the canvas origin-unclean
